pyproject.toml
[tool.poetry]
name = "automatik"
version = "0.1.0"
description = ""
authors = ["Axyss <contact@axyss.dev>"]
readme = "docs/README.md"

[tool.poetry.dependencies]
python = "^3.10"
discord-py = "^2.6.3"
requests = "^2.32.5"
beautifulsoup4 = "^4.13.5"
pyyaml = "^6.0.2"
pymongo = "^4.15.1"
dotenv = "^0.9.9"
curl-cffi = "^0.13.0"
google-genai = "^1.36.0"
agno = "^2.2.2"
igdb-api-v4 = "^0.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
